common

+++ src/main/common/math_robot.cpp
@@ -4,12 +4,12 @@
 #include <math.h>
 
 
-float m_rnd()
+double m_rnd()
 {
   return (rand()%200000000)/100000000.0 - 1.0;
 }
 
-float m_abs(float x)
+double m_abs(double x)
 {
   if (x < 0.0)
     x = -x;
@@ -17,7 +17,7 @@ float m_abs(float x)
   return x;
 }
 
-float m_sgn(float x)
+double m_sgn(double x)
 {
   if (x > 0.0)
     return 1.0;
@@ -28,7 +28,7 @@ float m_sgn(float x)
   return 0.0;
 }
 
-float m_saturate(float value_min, float value_max, float x)
+double m_saturate(double value_min, double value_max, double x)
 {
   if (x < value_min)
     x = value_min;
@@ -36,17 +36,17 @@ float m_saturate(float value_min, float value_max, float x)
   if (x > value_max)
     x = value_max;
 
-  return x;
+  return x; 
 }
 
-float m_map_to(float source_min, float source_max, float dest_min, float dest_max, float value)
+double m_map_to(double source_min, double source_max, double dest_min, double dest_max, double value)
 {
-  float k = (dest_max - dest_min)/(source_max - source_min);
-  float q = source_min - k*source_min;
+  double k = (dest_max - dest_min)/(source_max - source_min);
+  double q = source_min - k*source_min;
   return k*value + q;
 }
 
-float m_mix(float a, float b, float ratio)
+double m_mix(double a, double b, double ratio)
 {
   return ratio*a + (1.0 - ratio)*b;
 }
